добавил авторизацию

+++ fastapi_app/src/domain/comment/use_cases/delete_comment.py
@@ -1,6 +1,6 @@
 from src.infrastructure.sqlite.database import database
 from src.infrastructure.sqlite.repositories.comments import CommentRepository
-from src.core.exceptions import NotFoundError, DomainError, DatabaseError
+from src.core.exceptions import NotFoundError, AuthorizationError, DomainError, DatabaseError
 
 
 class DeleteComment:
@@ -8,7 +8,7 @@ class DeleteComment:
         self._database = database
         self._repo = CommentRepository()
 
-    async def execute(self, comment_id: int) -> bool:
+    async def execute(self, comment_id: int, current_user_id: int, is_superuser: bool) -> bool:
         try:
             with self._database.session() as session:
                 existing = self._repo.get_by_id(session, comment_id)
@@ -19,10 +19,13 @@ class DeleteComment:
                         value=str(comment_id)
                     )
                 
+                if existing.author_id != current_user_id and not is_superuser:
+                    raise AuthorizationError("Вы можете удалять только свои комментарии")
+                
                 deleted = self._repo.delete(session, comment_id)
                 return deleted
         
-        except NotFoundError:
+        except (NotFoundError, AuthorizationError):
             raise
         except DatabaseError as e:
             raise DomainError(
